# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoWVP Owl 的 HTTP API 文档，包含设备、通道、GB28181、PTZ、录像与系统接口。\n\n鉴权方式：\n大多数管理接口需要在请求头中携带 `Authorization: Bearer \u003ctoken\u003e`。\n可先调用 `/login/key` 获取 RSA 公钥，再调用 `/login` 完成登录。\n\n返回约定：\n1. 成功时通常返回业务对象，或 `{ \"msg\": \"ok\" }`\n2. 失败时通常返回 `{ \"msg\": \"错误原因\" }`\n3. GB28181 查询类接口通常会同时返回结构化结果与原始 XML",
        "title": "GoWVP Owl API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "使用前端 RSA 公钥加密后的用户名密码进行登录，返回 JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.loginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.loginOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/login/key": {
            "get": {
                "description": "获取用于登录前端加密的 RSA 公钥，返回 Base64 编码的 PEM 文本",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "获取登录公钥",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerLoginKeyOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/media_servers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "查询流媒体服务器列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "IP",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "类型",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerMediaServersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/media_servers/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "修改流媒体服务器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流媒体服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "流媒体服务器参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sms.EditMediaServerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sms.MediaServer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "修改登录凭据",
                "parameters": [
                    {
                        "description": "新用户名和密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateCredentialsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "api.SwaggerLoginKeyOutput": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "Base64 编码的 PEM 公钥，用于前端加密登录数据",
                    "type": "string"
                }
            }
        },
        "api.SwaggerMediaServersResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "流媒体服务器列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sms.MediaServer"
                    }
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "api.SwaggerMessageResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "处理结果说明",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "api.loginInput": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "Username string `json:\"username\" binding:\"required\"`\nPassword string `json:\"password\" binding:\"required\"`",
                    "type": "string",
                    "example": "Base64(RSA-OAEP(username/password JSON))"
                }
            }
        },
        "api.loginOutput": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "登录成功后返回的 JWT",
                    "type": "string"
                },
                "user": {
                    "description": "当前登录用户名",
                    "type": "string"
                }
            }
        },
...
        "api.updateCredentialsInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "新密码",
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "description": "新用户名",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "sms.EditMediaServerInput": {
            "type": "object",
            "properties": {
                "hook_ip": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "sdp_ip": {
                    "type": "string"
                },
                "secret": {
                    "description": "StreamIP          string           `json:\"stream_ip\"`\nPorts MediaServerPorts `json:\"ports\"`\nAutoConfig        bool             `json:\"auto_config\"`",
                    "type": "string"
                },
                "type": {
                    "description": "lalmax/zlm",
                    "type": "string"
                }
            }
        },
        "sms.MediaServer": {
            "type": "object",
            "properties": {
                "auto_config": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "hook_alive_interval": {
                    "type": "integer"
                },
                "hook_ip": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "last_keepalive_at": {
                    "type": "string"
                },
                "ports": {
                    "$ref": "#/definitions/sms.MediaServerPorts"
                },
                "record_assist_port": {
                    "type": "integer"
                },
                "record_day": {
                    "type": "integer"
                },
                "record_path": {
                    "type": "string"
                },
                "rtpenable": {
                    "type": "boolean"
                },
                "rtpport_range": {
                    "type": "string"
                },
                "sdp_ip": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "send_rtpport_range": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "stream_ip": {
                    "type": "string"
                },
                "transcode_suffix": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "sms.MediaServerPorts": {
            "type": "object",
            "properties": {
                "flv": {
                    "type": "integer"
                },
                "flvs": {
                    "type": "integer"
                },
                "http": {
                    "type": "integer"
                },
                "https": {
                    "type": "integer"
                },
                "rtmp": {
                    "type": "integer"
                },
                "rtmps": {
                    "type": "integer"
                },
                "rtpporxy": {
                    "type": "integer"
                },
                "rtsp": {
                    "type": "integer"
                },
                "rtsps": {
                    "type": "integer"
                },
                "ws_flv": {
                    "type": "integer"
                },
                "ws_flvs": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}