# internal/web/api/sms.go
// Code generated by godddx, DO AVOID EDIT.
package api

import (
	"github.com/gin-gonic/gin"
	"github.com/gowvp/owl/internal/conf"
	"github.com/gowvp/owl/internal/core/sms"
	"github.com/gowvp/owl/internal/core/sms/store/smsdb"
	"github.com/ixugo/goddd/pkg/orm"
	"github.com/ixugo/goddd/pkg/reason"
	"github.com/ixugo/goddd/pkg/web"
	"gorm.io/gorm"
)

type SmsAPI struct {
	smsCore sms.Core
	uc      *Usecase
}

func NewSMSCore(db *gorm.DB, cfg *conf.Bootstrap) sms.Core {
	core := sms.NewCore(smsdb.NewDB(db).AutoMigrate(orm.GetEnabledAutoMigrate()))
	if err := core.Run(cfg, cfg.Server.HTTP.Port); err != nil {
		panic(err)
	}
	return core
}

func NewSmsAPI(core sms.Core) SmsAPI {
	return SmsAPI{smsCore: core}
}

func registerSms(g gin.IRouter, api SmsAPI, handler ...gin.HandlerFunc) {
	{
		group := g.Group("/media_servers", handler...)
		group.GET("", web.WrapH(api.findMediaServer))
		group.PUT("/:id", web.WrapH(api.editMediaServer))

		// group.GET("/:id", web.WrapH(api.getMediaServer))
		// group.POST("", web.WrapH(api.addMediaServer))
		// group.DELETE("/:id", web.WrapH(api.delMediaServer))
	}
}

// >>> mediaServer >>>>>>>>>>>>>>>>>>>>

// findMediaServer godoc
// @Summary 查询流媒体服务器列表
// @Tags MediaServer
// @Security BearerAuth
// @Produce json
// @Param page query int false "页码"
// @Param size query int false "每页数量"
// @Param ip query string false "IP"
// @Param type query string false "类型"
// @Success 200 {object} SwaggerMediaServersResponse
// @Failure 400 {object} SwaggerErrorResponse
// @Router /media_servers [get]
func (a SmsAPI) findMediaServer(c *gin.Context, in *sms.FindMediaServerInput) (any, error) {
	items, total, err := a.smsCore.FindMediaServer(c.Request.Context(), in)
	return gin.H{"items": items, "total": total}, err
}

func (a SmsAPI) getMediaServer(c *gin.Context, _ *struct{}) (any, error) {
	mediaServerID := c.Param("id")
	return a.smsCore.GetMediaServer(c.Request.Context(), mediaServerID)
}

// editMediaServer godoc
// @Summary 修改流媒体服务器
// @Tags MediaServer
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "流媒体服务器ID"
// @Param body body sms.EditMediaServerInput true "流媒体服务器参数"
// @Success 200 {object} sms.MediaServer
// @Failure 400 {object} SwaggerErrorResponse
// @Router /media_servers/{id} [put]
func (a SmsAPI) editMediaServer(c *gin.Context, in *sms.EditMediaServerInput) (any, error) {
	mediaServerID := c.Param("id")
	out, err := a.smsCore.EditMediaServer(c.Request.Context(), in, mediaServerID, a.uc.Conf.Server.HTTP.Port)
	if err != nil {
		return nil, err
	}
	if mediaServerID == "local" {
		a.uc.Conf.Media.IP = out.IP
		a.uc.Conf.Media.SDPIP = out.SDPIP
		a.uc.Conf.Media.Secret = out.Secret
		a.uc.Conf.Media.WebHookIP = out.HookIP
		a.uc.Conf.Media.Type = out.Type
		if err := conf.WriteConfig(a.uc.Conf, a.uc.Conf.ConfigPath); err != nil {
			return nil, reason.ErrServer.SetMsg(err.Error())
		}
	}
	return out, err
}

func (a SmsAPI) addMediaServer(c *gin.Context, in *sms.AddMediaServerInput) (any, error) {
	return a.smsCore.AddMediaServer(c.Request.Context(), in)
}

func (a SmsAPI) delMediaServer(c *gin.Context, _ *struct{}) (any, error) {
	mediaServerID := c.Param("id")
	return a.smsCore.DelMediaServer(c.Request.Context(), mediaServerID)
}